requirements.txt
Flask==3.1.2
twilio==9.8.4
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.1.1
pdfplumber==0.11.7
pytesseract==0.3.13
pdf2image==1.17.0
pillow==12.0.0
google-generativeai==0.8.5
requests==2.32.5


